+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>pt.mystuff</groupId>
+    <groupId>pt.marcioneves</groupId>
     <artifactId>flow</artifactId>
     <version>1.0.0-RC01-SNAPSHOT</version>
     <packaging>jar</packaging>
