+++ src/main/java/pt/marcioneves/decision/core/AbstractContext.java
@@ -1,4 +1,4 @@
-package pt.mystuff.decision.core;
+package pt.marcioneves.decision.core;
 
 /**
  * This class represents the context to be passed across node navigation
