+++ src/test/java/pt/marcioneves/decision/FlowWorkbench.java
@@ -1,8 +1,8 @@
-package pt.mystuff.decision;
+package pt.marcioneves.decision;
 
 import org.junit.Test;
-import pt.mystuff.decision.core.AbstractContext;
-import pt.mystuff.decision.core.ICommand;
+import pt.marcioneves.decision.core.AbstractContext;
+import pt.marcioneves.decision.core.ICommand;
 
 import java.util.Random;
 import java.util.logging.Logger;
