+++ src/test/java/pt/marcioneves/decision/TwoNodesFlowTest.java
@@ -1,4 +1,4 @@
-package pt.mystuff.decision;
+package pt.marcioneves.decision;
 
 import org.junit.Assert;
 import org.junit.BeforeClass;
